requirements.txt
flask
validators
pytest
pyjwt